+++ resources/js/Pages/About/Interfaces/ChooseInterface.tsx
@@ -1,0 +1,13 @@
+interface AccordionItemData {
+  id: string
+  title: string
+  content: string
+}
+
+interface AccordionItemProps {
+  item: AccordionItemData
+  isOpen: boolean
+  onToggle: () => void
+}
+
+export type {AccordionItemData, AccordionItemProps}
