+++ resources/js/Pages/About/Interfaces/ExpertiseInterface.tsx
@@ -1,0 +1,6 @@
+
+export default interface ExpertiseCardProps {
+  icon: React.ComponentType<{ className?: string }>
+  title: string
+  description: string
+}
