+++ resources/js/Pages/About/Interfaces/PartnerInterface.tsx
@@ -1,0 +1,14 @@
+interface PartnerBenefit {
+  id: number
+  title: string
+  description: string
+  icon: React.ReactNode
+}
+
+interface PartnerSectionProps {
+  title?: string
+  subtitle?: string
+  benefits?: PartnerBenefit[]
+}
+
+export type { PartnerBenefit, PartnerSectionProps }
