+++ resources/js/Pages/About/Interfaces/TeamInterface.tsx
@@ -1,0 +1,12 @@
+interface TeamMember {
+  id: number
+  name: string
+  position: string
+  image: string
+}
+
+interface TeamCardProps {
+  member: TeamMember
+}
+
+export type {TeamMember, TeamCardProps}
